\name{entrezToGRanges}
\alias{entrezToGRanges}
\title{entrezToGRanges}
\usage{
  entrezToGRanges(entrezid)
}
\arguments{
  \item{entrezid}{character vector containing the ENTREZ
  IDs of the genes of interest.}
}
\value{
  and object of class \code{GenomicRanges}.
}
\description{
  Converts a set of ENTREZ IDs into genomic ranges (HG18)
}
\details{
  Converts a set of ENTREZ IDs into genomic ranges (HG18)
}
\author{
  ubuntu
}

