\docType{data}
\name{Segments}
\alias{Segments}
\title{Segmented data from the 304 CNV samples of the PETACC3 set}
\format{a data frame with 67097 rows and 6 columns}
\description{
  A dataset containing 67097 rows and 6 columns. Each row
  corresponds to a segment. The columns contein the
  following information: \itemize{ \item ID. The sample
  identifier (colnames of the CNV matrix) \item chrom. The
  chromosome where the segment is found. Note all the
  chromosomes are numeric (X = 23, Y = 24).  \item
  loc.start. Start position of the segment.  \item loc.end.
  End position of the segment.  \item num.mark. Number of
  markers falling in the segment.  \item seg.mean.
  Log-ratio mean value of the segment.
}
\keyword{datasets}

